perf(search): let the reductions table say it has no negative entry
`reductions[i]` is `int(2872/128.0 * ln i)` -- a scaled logarithm, 0 through 124
across the whole 256-entry table, and never negative. Stored as `i32` that range
is a fact no backend can use, so `reductionAcc`'s `@divTrunc(..., 512)` had to
allow for a negative dividend and paid the round-toward-zero correction on every
move the search reduces: the dividend materialised twice, one copy biased, the
sign tested, a cmov selecting. `u16` carries the range instead, the product and
the improving term are then provably non-negative, and the shift stands alone.

The widths are the bound and a test states it: 124 * 124 * 197 is 3,029,072, well
inside u32. `r[255] == 124` is checked over the 256 entries the worker actually
holds rather than the 64 the older test used -- the bound that matters is the one
at MAX_MOVES.

The table also halves, 1024 bytes to 512, inside the per-node scalar block
`WorkerLayout` places by hand; nothing here measures that.

Bit-exact: `zig build signature` OK -- bench == 2516158, and `zig build parity`'s
deterministic golden output is byte-identical to the pre-change run (diff = 0
lines outside `tb-search`'s node legs, `async`, `skill 0` and `time-mgmt`'s
wall-clock milliseconds).

MEASURED, x86-64-avx2, base md5 fc36dadb..., head md5 d5bfcbd3..., node totals
identical on every run:

  tools/ltc_ab.sh, warm 60-ply replay, depth 13, Hash 16, 3 rounds,
  1,141,939 nodes both sides
    Ir/node 8146.579 -> 8142.640   0.99952   (-3.94 per node)

  tools/perf_counters, cold `bench 16 1 13`, 5 rounds, 2,516,158 nodes
    Ir/node 7162 -> 7159

Small, and reported as such: the axis's A/A control on this column reads 1.00000
to five decimals, so -0.048% is above the floor, but it is a few instructions per
node and not a headline. No cycle figure is quoted -- the cycle and miss columns
read 1.089 and 1.121 on these two binaries, wider than the change, which is the
box.

  zig build test    354 tests passed, 4 skipped, exit 0
  zig fmt --check   clean

`zig build test` is what caught the retype's one loose end: a `fillReductions`
unit test still declared `[64]i32` and stood outside every root `parity` compiles.

+++ src/engine/search/search_ctx.zig
@@ -135,7 +135,7 @@ pub const QCtx = struct {
     nmp_min_ply: *i32,
     sel_depth: *i32,
     root_depth: *i32,
-    reductions: [*]const i32,
+    reductions: [*]const u16,
     root_delta: *const i32,
     last_iter_pv: *const PVMoves,
     stop: *const u8,
